requirements.txt
# Core scientific
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Geospatial
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.5.0
folium>=0.15.0
branca>=0.7.0

# Energy simulation
pandapipes>=0.8.0
pandapower>=2.13.0

# Machine learning
scikit-learn>=1.3.0

# LLM (optional)
google-genai>=0.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Economics module
pyyaml>=6.0
tqdm>=4.64.0
python-dotenv>=1.0.0
mypy>=1.8.0

# UHDC report rendering
jinja2>=3.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

